Move CityHash class to a separate file
